Updated and Added Database models
Did not implement connections to the game
