+++ stc_game/Assets/Resources/DevDogDb/MainLanguageDb.asset
@@ -12,7 +12,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 78b75a309ff767a4bb3ede0d5c1d2f6f, type: 3}
   m_Name: MainLanguageDb
   m_EditorClassIdentifier: 
-  _serializedJsonString: '{"_lang":"en-US","_localizedStrings":{"9d915740-ad6c-4ca5-890d-8218fdf76953":"Killed
+  _serializedJsonString: '{"_lang":"en-US","_localizedStrings":{"NO_KEY":"","9d915740-ad6c-4ca5-890d-8218fdf76953":"Killed
     {0} / {2} Goblin Chief","e7d0117b-9b70-4ddc-a094-deac664933de":"I feel as though
     I''m not ready to talk to you yet.","112184d4-5e92-4b2d-a926-4321250b1fd3":"Defeat
     Illamuwa","d6f17a10-22f7-4d52-8a11-918eb98758c1":"Nope","729f0368-c25a-43d9-8bf6-7f6cbe26b1c1":"Hello
