--- stc_game/Assets/Scripts/Models/AuthReq.cs
@@ -1,14 +1,0 @@
-﻿using System;
-
-[Serializable]
-public class AuthReq
-{
-    public string email;
-    public string password;
-
-    public AuthReq(string email, string password)
-    {
-        this.email = email;
-        this.password = password;
-    }
-}
