--- stc_game/Assets/Scripts/Models/AuthRes.cs
@@ -1,7 +1,0 @@
-﻿using System;
-
-[Serializable]
-public class AuthRes
-{
-    public string access_token;
-}
