--- stc_game/Assets/Scripts/Models/AuthRes.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 05764c1923cacc04bb6f6c785c76bcc0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
