--- stc_game/Assets/Scripts/Models/AuthUser.cs
@@ -1,15 +1,0 @@
-﻿using System;
-
-[Serializable]
-public class AuthUser
-{
-    public bool email_verified;
-    public string email;
-    public string updated_at;
-    public string name;
-    public string picture;
-    public string user_id;
-    public string nickname;
-    public string created_at;
-    public string sub;
-}
