--- stc_game/Assets/Scripts/Models/AuthUser.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6119c570a3415fb4fb66104c60b98c9a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
