--- stc_game/Assets/Scripts/Models/CharacterMage.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3e822c1f4030c9647bb362e4958e02d9
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
