+++ stc_game/Assets/Scripts/Models/DatabaseModels.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 86eb17382ab86e242aaa007ba989ae39
+guid: 97e1ddf155208bc4c922ca8eead01e27
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
