+++ stc_game/Assets/Scripts/Models/DatabaseModels/AuthenticationModels.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8c3fd2920bad13e4aa919146f16c124c
+guid: 8f4c3ec5b66f6634b8ec6f4d9a5e25a0
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
