+++ stc_game/Assets/Scripts/Models/DatabaseModels/AuthenticationModels/AuthReq.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+[Serializable]
+public class AuthReq
+{
+    public string email;
+    public string password;
+
+    public AuthReq(string email, string password)
+    {
+        this.email = email;
+        this.password = password;
+    }
+}
