+++ stc_game/Assets/Scripts/Models/DatabaseModels/AuthenticationModels/AuthReq.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5a2b2f4e49dcc394bae7d8b7a22e973b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
