+++ stc_game/Assets/Scripts/Models/DatabaseModels/AuthenticationModels/AuthRes.cs
@@ -1,0 +1,7 @@
+﻿using System;
+
+[Serializable]
+public class AuthRes
+{
+    public string access_token;
+}
