+++ stc_game/Assets/Scripts/Models/DatabaseModels/CharacterBaseModel.cs.meta
