+++ stc_game/Assets/Scripts/Models/DatabaseModels/CharacterKnightModel.cs.meta
