+++ stc_game/Assets/Scripts/Models/DatabaseModels/CharacterMage.cs.meta
