+++ stc_game/Assets/Scripts/Models/DatabaseModels/GameState.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+
+[Serializable]
+public class GameState
+{
+    public string UserId;
+    public string CharacterId;
+    public List<string> Abilities;
+}
