+++ stc_game/Assets/Scripts/Models/DatabaseModels/GameState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c9c2e8ff4bd84354fb9dce8f3b25c2a7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
