+++ stc_game/Assets/Scripts/Models/GameModels.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 14fe258e4bb602f4891309a8da98082b
+guid: 38fced6c964dfe244a883081e2e245df
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
