+++ stc_game/Assets/Scripts/Models/GameModels/GameCharacter.cs.meta
