+++ stc_game/Assets/Scripts/Models/GameModels/Hero.cs.meta
