--- stc_game/Assets/Scripts/Models/User.cs
@@ -1,15 +1,0 @@
-﻿using System;
-
-[Serializable]
-public class User: NewUser
-{
-    public string _id;
-
-    public User()
-    {
-        _id = "";
-        CharacterName = "";
-        UserAuthenticationId = "";
-        Character = new CharacterBaseModel();
-    }
-}
